requirements.txt
Flask
Flask-MySQLdb
wtforms
requests
mysql.connector
MySQLdb
